Update Team5Dprep.Rmd
I created the additional variable actorExperience which shows how much movies an actor has been in and also startyear_decades which groups the start years per decade. I also deleted the rows if the actor had been in less than 10 movies and if the movies had less than 50000 votes and kept only the top 5 genres the rest in other.